Add support for libp2p in nomos-node
